moves get_image out of Node, updates get_image unit tests.
